+++ src/components/Button/types.ts
@@ -1,0 +1,8 @@
+export interface CustomRedButtonProps  extends React.ButtonHTMLAttributes<HTMLButtonElement>{
+    title:string
+}
+
+export type ButtonDataType = {
+    id: number,
+    title: string,
+}
