+++ src/pages/ChatbotPage/types/index.ts
@@ -1,0 +1,9 @@
+export interface MessageProps {
+	profilePic: string | null | undefined;
+	message: string;
+}
+
+export interface UseChatScrollProps {
+    messages: MessageProps[],
+    streamMessage: MessageProps
+}
